+++ CHANGES.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Last release of this project is was 30th of September 2021.
 
+### Unreleased
+
+  - fix(ck5): setData understanding math LaTeX. #KB-39004
+
 ### 8.6.0 2023-10-10
 
   - Feat: Add branch name and commit hash to demo page